+++ packages/contracts/src/index.ts
@@ -1,4 +1,5 @@
 export * from './api'
+export * from './api-keys'
 export * from './assets'
 export * from './auth'
 export * from './canvas'